fix: bug when session is set by lookup command and logging
I changed the order of reading arguments handover by the lookup command. The 'session' attribute was not used, because before it was recognized, the plugin already fails because bw login returns error.

Now, the session is first set and then used when running the command.

Additionally I improved the error logging, when the bw command is not returning proper json. The error should is printed on the ansible run.